# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "support socks4, socks5, http proxy all in one",
        "title": "Mixed-Socks",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "v2.0.0"
    },
    "paths": {
        "/api/config": {
            "get": {
                "description": "get mixed socks config",
                "tags": [
                    "Config"
                ],
                "summary": "get",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.JSONResult"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.JSONResult"
                        }
                    }
                }
            },
            "post": {
                "description": "update mixed socks config",
                "tags": [
                    "Config"
                ],
                "summary": "save",
                "parameters": [
                    {
                        "description": "config",
                        "name": "scripts",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.Conf"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.JSONResult"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.JSONResult"
                        }
                    }
                }
            }
        },
        "/api/reload": {
            "post": {
                "description": "reload mixed socks config",
                "tags": [
                    "Operate"
                ],
                "summary": "reload",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.JSONResult"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.JSONResult"
                        }
                    }
                }
            }
        },
        "/api/start": {
            "post": {
                "description": "start mixed socks",
                "tags": [
                    "Operate"
                ],
                "summary": "start",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.JSONResult"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.JSONResult"
                        }
                    }
                }
            }
        },
        "/api/state": {
            "get": {
                "description": "get mixed socks state",
                "tags": [
                    "Operate"
                ],
                "summary": "state",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.JSONResult"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.JSONResult"
                        }
                    }
                }
            }
        },
        "/api/stop": {
            "post": {
                "description": "stop mixed socks",
                "tags": [
                    "Operate"
                ],
                "summary": "stop",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.JSONResult"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.JSONResult"
                        }
                    }
                }
            }
        },
        "/api/user": {
            "get": {
                "description": "List all user",
                "tags": [
                    "User"
                ],
                "summary": "list",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.JSONResult"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.JSONResult"
                        }
                    }
                }
            },
            "post": {
                "description": "create or update user",
                "tags": [
                    "User"
                ],
                "summary": "save",
                "parameters": [
                    {
                        "description": "user",
                        "name": "scripts",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.JSONResult"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.JSONResult"
                        }
                    }
                }
            },
            "delete": {
                "description": "delete user",
                "tags": [
                    "User"
                ],
                "summary": "delete",
                "parameters": [
                    {
                        "type": "string",
                        "description": "username",
                        "name": "username",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.JSONResult"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.JSONResult"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.JSONResult": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 0
                },
                "data": {},
                "message": {
                    "type": "string",
                    "example": "消息"
                }
            }
        },
        "service.Conf": {
            "type": "object",
            "properties": {
                "cidr": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "host": {
                    "type": "string"
                },
                "port": {
                    "type": "integer"
                }
            }
        },
        "service.User": {
            "type": "object",
            "required": [
                "username"
            ],
            "properties": {
                "cidr": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "password": {
                    "type": "string"
                },
                "remark": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    }
}